Allow the caller to force the recomputation of the version
The flag is added to permit users to always get the latest version.
This is important in scenarios where a server is monitoring a file
system and automatically reloading when the contents change. With
the force flag, the server can ensure it has the latest version
of the git hash.